requirements.txt
git+https://github.com/openai/CLIP.git
filelock==3.0.12
ipython==7.34.0
matplotlib==3.4.2
numexpr==2.8.4
numpy==1.23.5
opencv_python==4.5.5.64
opencv_python_headless==4.5.4.60
pandas==1.1.4
Pillow==8.2.0
PyYAML==5.4.1
Requests==2.30.0
torchmetrics==0.11.4
tqdm==4.64.1
wget==3.2
jsonargparse[signatures]>=4.17.0
lightning>=2.0.0
transformers
